refactor: Pagination 변수명 변경, 변수 추출, 공통컴포넌트 폴더로 이동
